Safer checks against integer overflows.
git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK8@1714690 13f79535-47bb-0310-9956-ffa450edef68

+++ core/sis-utility/src/main/java/org/apache/sis/internal/util/Numerics.java
@@ -180,7 +180,7 @@ public final class Numerics extends Static {
         if (data == null) return null;
         final int[] result = new int[data.length];
         for (int i=0; i<data.length; i++) {
-            result[i] = (int) Math.round(data[i]);
+            result[i] = Math.toIntExact(Math.round(data[i]));
         }
         return result;
     }
